+++ CHANGELOG.markdown
@@ -1,6 +1,9 @@
+next
+----
+* Added a `MINIMAL` pragma to `Adjunction`.
+
 4.3
 ---
-* Added a `MINIMAL` pragma to `Adjunction`.
 * Removed a spurious superclass constraint for `Applicative (StoreT g w)`
 * GHC 8 support
 * `comonad` 5 support