+++ CHANGELOG.markdown
@@ -9,6 +9,7 @@ next
 
 4.3
 ---
+* Added a `MINIMAL` pragma to `Adjunction`.
 * Removed a spurious superclass constraint for `Applicative (StoreT g w)`
 * GHC 8 support
 * `comonad` 5 support